% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosen_viz.R
\name{rosen_viz}
\alias{rosen_viz}
\title{Visualize Rosenbrock's Function in 3D using Plotly}
\usage{
rosen_viz(R = 0.825, cls = c("orange", "steelblue"), Rseq = 50, lseq = 100)
}
\arguments{
\item{R}{A numeric value specifying the maximum range of x and y values to visualize.}

\item{cls}{A vector of length 2 specifying the colors for the minimum and maximum values of the function (default: c('orange', 'steelblue')).}
}
\value{
A Plotly visualization of Rosenbrock's function in 3D.
}
\description{
This function generates a 3D visualization of Rosenbrock's function using Plotly. The function is visualized over a range of values for x and y.
}
\examples{
rosen_viz(R = 1)

}
