% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosenbrock.R
\name{rosenbrock}
\alias{rosenbrock}
\title{Compute the value of Rosenbrock's function for given parameters}
\usage{
rosenbrock(x, y, a = 1, b = 100)
}
\arguments{
\item{x}{A numeric value for the first parameter.}

\item{y}{A numeric value for the second parameter.}
}
\value{
The value of Rosenbrock's function for the given parameters.
}
\description{
Computes the value of the Rosenbrock's function for given parameters x and y
}
\examples{
rosenbrok(1, 1)

}
