% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosenopt_adam.R
\name{rosenopt_adam}
\alias{rosenopt_adam}
\title{Optimize the Rosenbrock's function using the ADAM algorithm}
\usage{
rosenopt_adam(
  init = c(5, 5),
  n.iter = 1000,
  alpha = 0.001,
  beta1 = 0.9,
  beta2 = 0.999,
  epsilon = 1e-08
)
}
\arguments{
\item{init}{A numeric vector containing the initial values of the parameters x and y (default: c(5, 5)).}

\item{n.iter}{A numeric value specifying the number of iterations to perform (default: 1e3).}

\item{alpha}{A numeric value specifying the initial step size.}

\item{beta1}{A numeric value specifying the exponential decay rate for the first moment estimates (default: 0.9).}

\item{beta2}{A numeric value specifying the exponential decay rate for the second moment estimates (default: 0.999).}

\item{epsilon}{A small positive numeric value as the threshold for the change in function value between iterations (default: 1e-8).#'}
}
\value{
A matrix with the value of the optimized Rosenbrock's parameters (x and y) and the function's output after each iteration. Each row represents an iteration, and the columns contain the x, y, and function output values, respectively.
}
\description{
Optimize the Rosenbrock's function using the ADAM (Adaptive Moment Estimation) algorithm.
}
\examples{
rosenopt_adam(init = c(5, 5), n.iter = 1e3, alpha = 1, epsilon = 1e-6)
}
