% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosenopt_bfgs.R
\name{rosenopt_bfgs}
\alias{rosenopt_bfgs}
\title{Optimize the Rosenbrock's function using the BFGS algorithm}
\usage{
rosenopt_bfgs(init = c(5, 5), n.iter = 1000, alpha = 1, epsilon = 1e-06)
}
\arguments{
\item{init}{A numeric vector with the initial values of the parameters x and y (default: c(5, 5)).}

\item{n.iter}{A numeric value specifying the number of iterations to perform (default: 1e3).}

\item{alpha}{A numeric value specifying the initial step size.}

\item{epsilon}{A small positive numeric value as the threshold for the change in function value between iterations (default: 1e-6).}
}
\value{
A matrix with the value of the optimized Rosenbrock's parameters (x and y) and the function's output after each iteration. Each row represents an iteration, and the columns contain the x, y, and function output values, respectively.
}
\description{
Optimize the Rosenbrock's function using the BFGS (Broyden-Fletcher-Goldfarb-Shanno) algorithm.
}
\examples{
rosenopt_bfgs(init = c(5, 5), n.iter = 1e3, alpha = 1, epsilon = 1e-6)
}
