% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosenopt_fr.R
\name{rosenopt_fr}
\alias{rosenopt_fr}
\title{Optimize the Rosenbrock's function using the Fletcher-Reeves algorithm}
\usage{
rosenopt_fr(init = c(5, 5), n.iter = 1000, alpha = 0.001)
}
\arguments{
\item{init}{A numeric vector with the initial values of the parameters x and y (default: c(5, 5)).}

\item{n.iter}{A numeric value specifying the number of iterations to perform (default: 1e3).}

\item{alpha}{A numeric value specifying the initial learning rate.}
}
\value{
A matrix with the value of the optimized Rosenbrock's parameters (x and y) and the functions output after each iteration on each column.
}
\description{
Optimize the Rosenbrock's function using the Fletcher-Reeves algorithm.
}
\examples{
rosenopt_fr(n.iter=1e5)

}
