% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosenopt_shc.R
\name{rosenopt_shc}
\alias{rosenopt_shc}
\title{Optimize the Rosenbrock's function using a stochastic hill climbing algorithm}
\usage{
rosenopt_shc(
  init = c(5, 5),
  n.iter = 1000,
  alpha = 0.5,
  batch = 32,
  sigma = 0.01
)
}
\arguments{
\item{init}{A numeric vector with the initial values of the parameters x and y (default: c(5, 5)).}

\item{n.iter}{A numeric value specifying the number of iterations to perform (default: 1e3).}

\item{alpha}{A numeric value specifying the step size.}

\item{batch}{An integer value specifying the number of candidates to generate at each iteration.}

\item{sigma}{A numeric value specifying the standard deviation for generating candidates.}
}
\value{
A matrix with the value of the optimized Rosenbrock's parameters (x and y) and the function's output after each iteration. Each row represents an iteration, and the columns contain the x, y, and function output values, respectively.
}
\description{
Optimize the Rosenbrock's function using a stochastic hill climbing algorithm.
}
\examples{
rosenopt_shc(init = c(5, 5), n.iter = 1e3, alpha = 1e-3, batch = 32, sigma = 1e-6)
}
