% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosen_vizopt.R
\name{rosen_vizopt}
\alias{rosen_vizopt}
\title{Visualize Optimization trajectroy for Rosenbrock's Function}
\usage{
rosen_vizopt(
  R = 3,
  cls = c("orange", "steelblue", "red", "navy"),
  optimizer_output,
  psy = 100,
  main,
  pnts = TRUE,
  mag = 1.5
)
}
\arguments{
\item{R}{A numeric value specifying the maximum range of x and y values to visualize.}

\item{cls}{A vector of length 3 specifying the colors for the minimum and maximum values of the function and the trajectory (default: c('orange', 'steelblue','red')).}

\item{psy}{A positive integer value specifying how many rows of the optimizer_output to skip before adding a point to the plot (default: 100).}
}
\value{
A Plotly visualization in 3D.
}
\description{
Visualize Optimization trajectroy for Rosenbrock's Function
}
\examples{
data = rosenopt_gd(n.iter=1e6,init=c(2,2))
rosen_vizopt(optimizer_output=data,psy=1000,R=1)

}
